Refactor account-related models, views, and templates, and update URLs
