+++ organizer_py/urls.py
@@ -24,5 +24,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("users/", include("users.urls")),
     path("users/", include("django.contrib.auth.urls")),
-    path("organizer/", include("organizer.urls")),
+    path("account/", include("account.urls")),
 ]
