+++ templates/accounts/account_form.html
@@ -15,7 +15,7 @@
         <input type="submit" class="btn btn-primary btn-lg" value="Submit"
           style="margin-right: 30px; margin-top: 30px;">
         <input type="button" class="btn btn-danger btn-lg" style="margin-top: 30px;"
-          onclick="window.location.href='{% url 'organizer:accounts_list' %}';" value="Cancel">
+          onclick="window.location.href='{% url 'account:accounts_list' %}';" value="Cancel">
       </div>
     </form>
