+++ templates/accounts/account_list.html
@@ -5,7 +5,7 @@
 {% block content %}
 
 <h2>Accounts List</h2>
-<a href="{% url 'organizer:create_account' %}">Cadastrar Conta</a>
+<a href="{% url 'account:create_account' %}">Cadastrar Conta</a>
 
 <ul>
   {% for account in accounts %}
@@ -14,12 +14,12 @@
     {{ account.balance_cents }}
     {{ account.kind }}
     <div style="display: inline-block;">
-      <a href="{% url 'organizer:update_account' account.id %}">
+      <a href="{% url 'account:update_account' account.id %}">
         <button type="button">Update</button>
       </a>
     </div>
     <div style="display: inline-block;">
-      <form method="post" action="{% url 'organizer:delete_account' account.id %}">
+      <form method="post" action="{% url 'account:delete_account' account.id %}">
         {% csrf_token %}
         <input type="submit" value="Delete">
       </form>
