+++ users/models.py
@@ -1,8 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
 
-# from organizer.models import Account
-
 
 class Category(models.Model):
     name = models.CharField(max_length=100)
@@ -52,10 +50,10 @@ class Installment(models.Model):
 
 class Transference(models.Model):
     receiver = models.ForeignKey(
-        "organizer.Account", related_name="receiver", on_delete=models.CASCADE
+        "account.Account", related_name="receiver", on_delete=models.CASCADE
     )
     sender = models.ForeignKey(
-        "organizer.Account", related_name="sender", on_delete=models.CASCADE
+        "account.Account", related_name="sender", on_delete=models.CASCADE
     )
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     value_cents = models.IntegerField()
